Fix gi list file size overflow, JIRA:SB-2519
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@85867 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/blast/seqdb_reader/seqdbgilistset.cpp
@@ -96,7 +96,7 @@ public:
         
         const char * fbeginp = memlease.GetFileDataPtr(0);
         
-        const char * fendp   = fbeginp + (int)file_size;
+        const char * fendp   = fbeginp + file_size;
 
         try {
             bool in_order = false;
